Removing dependencies between decidim-admin and other engines
