+++ decidim-admin/app/forms/decidim/admin/participatory_process_user_role_form.rb
@@ -14,10 +14,10 @@ module Decidim
 
       validates :email, :role, presence: true
       validates :name, presence: true
-      validates :role, inclusion: { in: ParticipatoryProcessUserRole::ROLES }
+      validates :role, inclusion: { in: Decidim::ParticipatoryProcessUserRole::ROLES }
 
       def roles
-        Decidim::Admin::ParticipatoryProcessUserRole::ROLES.map do |role|
+        Decidim::ParticipatoryProcessUserRole::ROLES.map do |role|
           [
             I18n.t(role, scope: "decidim.admin.models.participatory_process_user_role.roles"),
             role
