+++ decidim-admin/app/models/decidim/admin/abilities/admin_user.rb
@@ -9,7 +9,7 @@ module Decidim
         include CanCan::Ability
 
         def initialize(user, _context)
-          return unless user && user.role?(:admin)
+          return unless user && user.admin?
 
           can :manage, ParticipatoryProcess
           can :manage, ParticipatoryProcessGroup
