+++ decidim-admin/app/views/decidim/admin/participatory_process_user_roles/index.html.erb
@@ -2,7 +2,7 @@
   <div class="card-divider">
     <h2 class="card-title">
       <%= t(".process_admins_title", scope: "decidim.admin") %>
-      <% if can? :create, Decidim::Admin::ParticipatoryProcessUserRole %>
+      <% if can? :create, Decidim::ParticipatoryProcessUserRole %>
          <%= link_to t("actions.new", scope: "decidim.admin", name: t("models.participatory_process_user_role.name", scope: "decidim.admin")), new_participatory_process_user_role_path(participatory_process), class: 'button tiny button--title new' %>
       <% end %>
     </h2>
