+++ decidim-admin/lib/decidim/admin/test/factories.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 FactoryGirl.define do
-  factory :participatory_process_user_role, class: Decidim::Admin::ParticipatoryProcessUserRole do
+  factory :participatory_process_user_role, class: Decidim::ParticipatoryProcessUserRole do
     user
     participatory_process
     role "admin"
