+++ decidim-admin/spec/queries/process_admins_spec.rb
@@ -8,7 +8,7 @@ describe Decidim::Admin::ProcessAdmins do
   let!(:admin) { create(:user, :admin, :confirmed, organization: organization) }
   let!(:participatory_process_admin) do
     user = create(:user, :confirmed, organization: organization)
-    Decidim::Admin::ParticipatoryProcessUserRole.create!(
+    Decidim::ParticipatoryProcessUserRole.create!(
       role: :admin,
       user: user,
       participatory_process: participatory_process
