+++ decidim-core/app/commands/decidim/remove_admin.rb
@@ -1,27 +1,25 @@
 # frozen_string_literal: true
 
 module Decidim
-  # A command to remove a role from a given User.
-  class RemoveUserRole < Rectify::Command
+  # A command to remove the admin privilege to an user.
+  class RemoveAdmin < Rectify::Command
     # Public: Initializes the command.
     #
     # form - A form object with the params.
-    def initialize(user, role)
+    def initialize(user)
       @user = user
-      @role = role
     end
 
     def call
       return broadcast(:invalid) unless user
 
-      user.roles.delete(role.to_s)
-      user.save!
+      user.update_attribute!(:admin, false)
 
       broadcast(:ok)
     end
 
     private
 
-    attr_reader :user, :role
+    attr_reader :user
   end
 end
